Moved files around and combined what had to be combined.
WebComponentExporter is now an abstract class. Test have not been fixed and the
initialization process that is in WebComponentExporterRegistryInitializer needs
to be transferred into the correct place.
