+++ flow-server/src/main/java/com/vaadin/flow/component/webcomponent/PropertyConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2000-2018 Vaadin Ltd.
+ * Copyright 2000-2019 Vaadin Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -17,19 +17,46 @@
 package com.vaadin.flow.component.webcomponent;
 
 import java.io.Serializable;
+import java.util.Objects;
 
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.function.SerializableBiConsumer;
 
 /**
- * Interface used to configure a single property on a web component created
- * by subclassing {@link com.vaadin.flow.component.WebComponentExporter}.
- * Offers a fluent API for configuring the property.
+ * Offers a fluent API for configuring the properties of embedded web
+ * components produced by
+ * {@link com.vaadin.flow.component.WebComponentExporter}.
  *
  * @param <C>   type of the {@code component} exported as a web component
  * @param <P>   type of the property exposed on the web component
  */
-public interface PropertyConfiguration<C extends Component, P extends Serializable> extends Serializable {
+public abstract class PropertyConfiguration<C extends Component,
+        P extends Serializable> {
+    private Class<C> componentClass;
+    private PropertyData<P> data;
+    private SerializableBiConsumer<C, Serializable> onChangeHandler = null;
+
+    private PropertyConfiguration() {}
+
+    /**
+     * Constructs a new {@code PropertyConfigurationImpl} tied to the
+     * exported {@link Component} type given by {@code componentType}.
+     *
+     * @param componentType     type of the exported {@code component}
+     * @param propertyName      name of the property
+     * @param propertyType      type of the property
+     * @param defaultValue      default value of the property. If the
+     *                          property type has a primitive version, this
+     *                          value is used when ever the property is being
+     *                          set to a {@code null}.
+     */
+    protected PropertyConfiguration(Class<C> componentType, String propertyName,
+                                 Class<P> propertyType, P defaultValue) {
+
+        data = new PropertyData<>(propertyName, propertyType, false,
+                defaultValue);
+        this.componentClass = componentType;
+    }
 
     /**
      * Sets a Property change handler. {@code onChange} can only be called
@@ -65,13 +92,59 @@ public interface PropertyConfiguration<C extends Component, P extends Serializab
      *                          the property value
      * @return this {@code PropertyConfiguration}
      */
-    PropertyConfiguration<C, P> onChange(
-            SerializableBiConsumer<C, P> onChangeHandler);
+    public PropertyConfiguration<C, P> onChange(SerializableBiConsumer<C, P> onChangeHandler) {
+        Objects.requireNonNull(onChangeHandler, "Parameter 'onChangeHandler' " +
+                "cannot be null!");
+        if (this.onChangeHandler != null) {
+            throw new IllegalStateException(String.format("onChangeHandler " +
+                    "for property %s has already been set and cannot be " +
+                    "overwritten!", data.getName()));
+        }
+        this.onChangeHandler = (c, o) -> onChangeHandler.accept(c, (P) o);
+        return this;
+    }
 
     /**
      * Mark the property as read-only. It cannot be written to by the client.
      *
      * @return this {@code PropertyConfiguration}
      */
-    PropertyConfiguration<C, P> readOnly();
+    public PropertyConfiguration<C, P> readOnly() {
+        data = new PropertyData<>(data, true);
+        return this;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj instanceof PropertyConfiguration) {
+            PropertyConfiguration other = (PropertyConfiguration) obj;
+            return data.equals(other.data)
+                    && componentClass.equals(other.componentClass);
+        }
+        return false;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(componentClass, data);
+    }
+
+    /**
+     * Retrieves the {@code onChangeHandler} tied to this property, if on
+     * exists.
+     *
+     * @return handler or {@code null}
+     */
+    protected SerializableBiConsumer<C, Serializable> getOnChangeHandler() {
+        return onChangeHandler;
+    }
+
+    /**
+     * Computed {@link PropertyData} based on the configuration details.
+     *
+     * @return {@code PropertyData} value object
+     */
+    protected PropertyData<P> getPropertyData() {
+        return data;
+    }
 }
