+++ flow-server/src/main/java/com/vaadin/flow/component/webcomponent/PropertyData.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2000-2018 Vaadin Ltd.
+ * Copyright 2000-2019 Vaadin Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.server.webcomponent;
+package com.vaadin.flow.component.webcomponent;
 
 import java.io.Serializable;
 import java.util.Objects;
