+++ flow-server/src/main/java/com/vaadin/flow/server/communication/WebComponentBootstrapHandler.java
@@ -28,7 +28,7 @@ import com.vaadin.flow.server.VaadinResponse;
 import com.vaadin.flow.server.VaadinServletRequest;
 import com.vaadin.flow.server.VaadinServletService;
 import com.vaadin.flow.server.VaadinSession;
-import com.vaadin.flow.server.webcomponent.WebComponentConfigurationRegistry;
+import com.vaadin.flow.server.webcomponent.WebComponentExporterRegistry;
 import com.vaadin.flow.shared.ApplicationConstants;
 
 import elemental.json.JsonObject;
@@ -52,7 +52,7 @@ public class WebComponentBootstrapHandler extends BootstrapHandler {
                 Class<T> annotationType) {
             ServletContext servletContext = ((VaadinServletService) getRequest()
                     .getService()).getServlet().getServletContext();
-            WebComponentConfigurationRegistry registry = WebComponentConfigurationRegistry
+            WebComponentExporterRegistry registry = WebComponentExporterRegistry
                     .getInstance(servletContext);
             return registry.getEmbeddedApplicationAnnotation(annotationType);
         }
