+++ flow-server/src/main/java/com/vaadin/flow/server/webcomponent/OSGiWebComponentExporterRegistry.java
@@ -24,7 +24,7 @@ import com.vaadin.flow.component.WebComponentExporter;
  * Data collector component for collecting web components in an OSGi
  * environment.
  */
-public class OSGiWebComponentConfigurationRegistry extends WebComponentConfigurationRegistry {
+public class OSGiWebComponentExporterRegistry extends WebComponentExporterRegistry {
 
     /**
      * {@inheritDoc}
